// src/lib/enums/product.enum.ts
export enum ProductMaterial {
  GOLD = "GOLD",
  SILVER = "SILVER",
  PLATINUM = "PLATINUM",
}

export enum ProductSize {
  XS = "XS",
  S = "S",
  M = "M",
  L = "L",
  XL = "XL",
  XXL = "2XL",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  RING = "RING",
  NECKLACE = "NECKLACE",
  BRACELET = "BRACELET",
  WOMANS_WATCHES = "WOMANS_WATCHES",
  EARRING = "EARRING",
}